[ ♻️  | | refactor ] 장바구니 담기 api 관련 로직 수정
2024.09.10

+++ src/main/java/pintoss/giftmall/domains/cart/service/CartService.java
@@ -11,7 +11,9 @@ import pintoss.giftmall.domains.cart.infra.CartRepository;
 import pintoss.giftmall.domains.cart.infra.CartReader;
 import pintoss.giftmall.domains.product.domain.PriceCategory;
 import pintoss.giftmall.domains.product.domain.Product;
+import pintoss.giftmall.domains.product.domain.ProductImage;
 import pintoss.giftmall.domains.product.infra.PriceCategoryReader;
+import pintoss.giftmall.domains.product.infra.ProductImageRepository;
 import pintoss.giftmall.domains.product.infra.ProductReader;
 import pintoss.giftmall.domains.user.domain.User;
 import pintoss.giftmall.domains.user.infra.UserReader;
@@ -31,12 +33,12 @@ public class CartService {
     private final ProductReader productReader;
     private final UserReader userReader;
     private final PriceCategoryReader priceCategoryReader;
+    private final ProductImageRepository productImageRepository;
 
     public Long addToCart(Long userId, Long productId, CartRequest requestDTO) {
         Product product = productReader.findById(productId);
         PriceCategory priceCategory = priceCategoryReader.findById(requestDTO.getPriceCategoryId());
         User user = userReader.findById(userId);
-
 
         Optional<Cart> existingCartOptional = cartRepository.findByUserAndProductAndPriceCategory(user, product, priceCategory);
 
@@ -52,7 +54,6 @@ public class CartService {
         return cart.getId();
     }
 
-
     @Transactional(readOnly = true)
     public List<CartResponse> getCartItems(Long userId) {
         userReader.findById(userId);
@@ -60,25 +61,32 @@ public class CartService {
         List<Cart> cartItems = cartRepository.findAllByUserId(userId);
 
         return cartItems.stream()
-                .map(CartResponse::fromEntity)
+                .map(cart -> {
+                    String logoImageUrl = productImageRepository.findByProductId(cart.getProduct().getId())
+                            .stream()
+                            .findFirst()
+                            .map(ProductImage::getUrl)
+                            .orElse(null);
+                    return CartResponse.fromEntity(cart, logoImageUrl);
+                })
                 .collect(Collectors.toList());
     }
 
-
     public void updateCartItem(Long cartItemId, int quantity) {
         Cart cart = cartReader.findById(cartItemId);
-
         cart.updateQuantity(quantity);
     }
 
     public void deleteCartItem(Long cartItemId) {
         Cart cart = cartReader.findById(cartItemId);
-
         cartRepository.delete(cart);
     }
 
     public void deleteAllCartItems(Long userId) {
         List<Cart> cartItems = cartRepository.findAllByUserId(userId);
+        if (cartItems.isEmpty()) {
+            throw new IllegalArgumentException("장바구니가 비어 있습니다.");
+        }
         cartRepository.deleteAll(cartItems);
     }
 
@@ -90,5 +98,5 @@ public class CartService {
         }
     }
 
-
 }
+
